Allow TempDir::remove()

+++ tempfile/TempDir.cpp
@@ -5,12 +5,19 @@ namespace bf = boost::filesystem;
 namespace cpputils {
 
 TempDir::TempDir()
-  : _path(bf::unique_path(bf::temp_directory_path() / "%%%%-%%%%-%%%%-%%%%")) {
+  : _path(bf::unique_path(bf::temp_directory_path() / "%%%%-%%%%-%%%%-%%%%")), _existing(true) {
   bf::create_directory(_path);
 }
 
 TempDir::~TempDir() {
-  bf::remove_all(_path);
+  remove();
+}
+
+void TempDir::remove() {
+  if (_existing) {
+    bf::remove_all(_path);
+    _existing = false;
+  }
 }
 
 const bf::path &TempDir::path() const {
